Refactor swifttool
This change simply moves SwiftRingsDefinition to its own file. The code
was becoming too unweildy in a single file.
